+++ packages/studio/src/hooks/useServerConnection.ts
@@ -1,0 +1,71 @@
+import { useState } from "react";
+import { buildProjectHash, parseProjectIdFromHash } from "../utils/projectRouting";
+import { useMountEffect } from "./useMountEffect";
+
+interface ServerConnectionState {
+  projectId: string | null;
+  resolving: boolean;
+  waitingForServer: boolean;
+}
+
+/**
+ * Resolves the active project ID by pinging /api/projects.
+ *
+ * If the hash contains a project ID the server is still contacted — this
+ * ensures a dead server (bookmark-reload case) enters the waiting state
+ * rather than mounting the full Studio against a non-responsive API.
+ *
+ * Polls every 2 s until the server responds, then transitions automatically.
+ * Cleans up pending timers on unmount so it is safe under React StrictMode.
+ */
+export function useServerConnection(): ServerConnectionState {
+  const [projectId, setProjectId] = useState<string | null>(null);
+  const [resolving, setResolving] = useState(true);
+  const [waitingForServer, setWaitingForServer] = useState(false);
+
+  useMountEffect(() => {
+    const hashProjectId = parseProjectIdFromHash(window.location.hash);
+    let cancelled = false;
+    let retryTimer: ReturnType<typeof window.setTimeout> | null = null;
+
+    function scheduleRetry() {
+      setWaitingForServer(true);
+      retryTimer = window.setTimeout(tryConnect, 2000);
+    }
+
+    function tryConnect() {
+      fetch("/api/projects")
+        .then((r) => r.json())
+        .then((data) => {
+          if (cancelled) return;
+          if (hashProjectId) {
+            setProjectId(hashProjectId);
+            setWaitingForServer(false);
+          } else {
+            const first = (data.projects ?? [])[0];
+            if (first) {
+              setProjectId(first.id);
+              setWaitingForServer(false);
+              window.location.hash = buildProjectHash(first.id);
+            } else {
+              scheduleRetry();
+            }
+          }
+        })
+        .catch(() => {
+          if (!cancelled) scheduleRetry();
+        })
+        .finally(() => {
+          if (!cancelled) setResolving(false);
+        });
+    }
+
+    tryConnect();
+    return () => {
+      cancelled = true;
+      if (retryTimer !== null) clearTimeout(retryTimer);
+    };
+  });
+
+  return { projectId, resolving, waitingForServer };
+}
